fix(store): pass uint32 lock flags to Windows LockFileEx
TryLock inferred int from the untyped LOCKFILE_EXCLUSIVE_LOCK constant,
so the Windows build failed. Cast the flags to uint32.

+++ internal/securefile/lock_windows.go
@@ -41,7 +41,7 @@ func lock(path string, disposition uint32, nonBlocking bool) (func() error, erro
 		file.Close()
 		return nil, err
 	}
-	lockFlags := windows.LOCKFILE_EXCLUSIVE_LOCK
+	lockFlags := uint32(windows.LOCKFILE_EXCLUSIVE_LOCK)
 	if nonBlocking {
 		lockFlags |= windows.LOCKFILE_FAIL_IMMEDIATELY
 	}
